backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
